Control startup order of docker-compose containers
Based on a healthcheck the mysql container starts first and the web app
waits for mysql to be ready to receive connections.
The only problem is that when running locally for the first time and
there is no prior mysql volume, the mysql server reinitializes and is
too slow. Thus the web app fails, but upon restart it works correctly.